+++ CHANGES.md
@@ -1,6 +1,34 @@
 
 # Unreleased
 
+# 9.0.0b1 
+
+* change: `assets` signature in `MultiBaseReader` method to be a union of `Sequence[str | dict] | str | dict | none` 
+* remove: `asset` parsing in `STACReader._get_asset_info()` method. Options needs to be passed as dictionary:
+    ```python
+    # 9.0.0a1
+    with STACReader(stac.json) as stac:
+        _ = stac.tile(
+            0, 
+            0, 
+            0, 
+            assets=[
+                "asset|indexes=1,2,3",
+            ]
+        )
+
+    # 9.0.0b1
+    with STACReader(stac.json) as stac:
+        _ = stac.tile(
+            0, 
+            0, 
+            0, 
+            assets=[
+                {"name": "asset", "indexes": [1, 2, 3]},
+            ]
+        )
+    ```
+
 # 9.0.0a6 (2026-02-13)
 
 * add: wraps readers within the mosaic Backends with `inherit_rasterio_env` decorator.